day 06 part 1 test data ok

+++ utils.c
@@ -11,7 +11,7 @@ inline Point add_points(const Point *point1, const Point *point2) {
     return (Point){point1->x + point2->x, point1->y + point2->y};
 }
 
-int reallocate_int_array(int **array, const int new_count) {
+int reallocate_int_array(int **array, const size_t new_count) {
     int *new_array = realloc(*array, new_count * sizeof(int *));
     if (new_array == NULL) {
         perror("Out of memory");
@@ -21,6 +21,7 @@ int reallocate_int_array(int **array, const int new_count) {
     return 1;
 }
 
+// TODO should be removed
 int reallocate_int_jagged_array(int ***array, const int new_count) {
     int **new_array = realloc(*array, new_count * sizeof(int *));
     if (new_array == NULL) {
